Break VoxelFS.glsl into smaller components
